Re-create linalg.h header after it has been renamed to coopvec.h
